Moved TileReader into it's own header and source file.
